Minor changes and renames
